+++ src/violet/conf/Config.hpp
@@ -39,17 +39,19 @@ struct Config {
                 "The prefix must start with a /, or be an empty string, but found \"{}\"",
                 newPrefix
             );
-            throw std::runtime_error(
-                "Illegal prefix"
-            );
+            throw std::runtime_error("Illegal prefix");
         } else if (!newPrefix.empty() && newPrefix.ends_with("/")) {
             minilog::error(
                 "The prefix must not end with a / (unless the entire prefix is \"/\"), but found \"{}\"",
                 newPrefix
             );
-            throw std::runtime_error(
-                "Illegal prefix"
-            );
+            throw std::runtime_error("Illegal prefix");
+        } else if (newPrefix.contains("//")) {
+            minilog::error("// is not allowed in the prefix");
+            throw std::runtime_error("Illegal prefix");
+        } else if (newPrefix.contains("/../") || newPrefix.ends_with("/..")) {
+            minilog::error("Path traversal is not allowed in the prefix");
+            throw std::runtime_error("Illegal prefix");
         } else {
             prefix = newPrefix;
         }